Force deployment so that FUNCTIONS_EXTENSION_VERSION is updated
This avoids having to set the version explicitly. If not forced, this error is shown:

```
You're trying to use v4 tooling to publish to a non-v4 function app (FUNCTIONS_EXTENSION_VERSION is set to ~3).
You can pass --force to force update the app to v4, or downgrade tooling for publishing.
```